% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adabag.R
\name{Adabag}
\alias{Adabag}
\title{Bagging classifier}
\usage{
Adabag(data, mfinal.seq)
}
\arguments{
\item{data}{A list generated by the function PrepareData. It is a list with two components. 
One is a vector of labels, the other is a data frame of features. The other element of the list
is a data frame of test features.}

\item{mfinal.seq}{A vector of values of number of iterations that should be tested.}
}
\value{
A list with a dataframe of errors a vector of predictions with the best parameters
}
\description{
Classify the input with a Bagging classifier. Function uses a vector of number of iterations
to search for the best training error. Once the best parameters have been identified, it uses them 
to predict on the test data.
}
\examples{
path <- "/home/rishabh/mres/ml_comp/data/"
data <- PrepareData(path, mode = 2, sample = TRUE, size = 100)
mfinal.seq <- seq(20, 25, 5)
Adabag(data, mfinal.seq)
}

