% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNN1.R
\name{KNN1}
\alias{KNN1}
\title{kNN classifier}
\usage{
KNN1(data, k.seq)
}
\arguments{
\item{data}{A list generated by the function PrepareData. It is a list with two components. 
One is a vector of labels, the other is a data frame of features. The other element of the list
is a data frame of test features.}

\item{k.seq}{A vector vector of vaues of k that should be tested.}
}
\value{
A list with a dataframe of errors with corresponding k and a vector of predictions with
the best k
}
\description{
Classify the input with a k nearest neighbors classifier.
}
\examples{
# create dataset
path <- "/home/rishabh/mres/ml_comp/data/"
data <- PrepareData(path, mode = 2, sample = TRUE, size = 100)
k.seq <- seq(1, 15, 2)
knn.result <- KNN1(data, k.seq)
}

