% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareData.R
\name{Normalise}
\alias{Normalise}
\title{Normalises data}
\usage{
Normalise(vector, vec.mean = NA, vec.sd = NA)
}
\arguments{
\item{vector}{A feature}

\item{vec.mean}{Is the sample mean if provided}

\item{vec.sd}{Is the sample sd if provided}
}
\value{
A list of normalised vector and sample stats
}
\description{
Normalises data
}
\examples{
vec <- Normalise(rnorm(100), 0.1, 1.1)
}

