% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareData.R
\name{PrepareData}
\alias{PrepareData}
\title{Preprocess Data}
\usage{
PrepareData(path, mode = 0, sample = FALSE, size = 1000)
}
\arguments{
\item{path}{provides path to training and test data}

\item{mode}{Defines the kind of preprocessing to be done. mode 0 does not process any data but seperates features from labels and returns 
a training set and a test set. mode1 creates z scores out of all the continuous features in training data set
For test data set the training sample means and sample variances are used to normalize data. 
mode 2 resizes all continuous features between 0 and 1}

\item{sample}{If sample is set to TRUE only a subset of processed data is returned}

\item{size}{The size of subset of data to be returned of \code{sample} set to TRUE}
}
\value{
A list with two components.
train - A list with two components. One is a vector of labels, the other is a data frame of features.
test - If mode is not 1 A data frame of features. Else a list where one component is the sample mean 
and sample sd used to normalize data and the other component is the list of test features
}
\description{
Preprocesses data based on input
}
\examples{
path <- "/home/rishabh/mres/ml_comp/data/"
data <- PrepareData(path, mode = 2, sample = TRUE, size = 100)
}

