% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareData.R
\name{Resize}
\alias{Resize}
\title{Resize vector}
\usage{
Resize(vector, vec.min, vec.max)
}
\arguments{
\item{vector}{A feature}

\item{vec.min}{Minimum value of feature}

\item{vec.max}{Maximum value of feature}
}
\value{
Resized vector
}
\description{
Resize vector
}
\examples{
Resize(rnorm(100, 5, 1), 2, 6)  
}

