% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareData.R
\name{ResizeFeatures}
\alias{ResizeFeatures}
\title{Resize Features}
\usage{
ResizeFeatures(train.feat, test.feat)
}
\arguments{
\item{train.feat}{A training data frame}

\item{test.feat}{A test data frame}
}
\value{
list of training and test features
normalises continous features between 0 and 1
}
\description{
Resize Features
}

