% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RF1.R
\name{RF1}
\alias{RF1}
\title{Random Forest classifier}
\usage{
RF1(data, ntree.vec)
}
\arguments{
\item{data}{A list generated by the function PrepareData. It is a list with two components. 
One is a vector of labels, the other is a data frame of features. The other element of the list
is a data frame of test features.}

\item{ntree.vec}{A vector vector of vaues of number of trees that should be tested.}
}
\value{
A list with a dataframe of errors a vector of predictions with the best parameters
}
\description{
Classify the input with a Random forest classifier. Function uses a vector of tree depth
and does a grid search with mtry parameter for the best training error. Once the best
Parameters have been identified, it uses them to predict on the test data.
}
\examples{
path <- "/home/rishabh/mres/ml_comp/data/"
data <- PrepareData(path, mode = 2, sample = TRUE, size = 100)
ntree.vec <- seq(50, 100, 10)
RF1(data, ntree.vec)
}

