% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareData.R
\name{SepFeatures}
\alias{SepFeatures}
\title{Seperate Features into continuous and categorical}
\usage{
SepFeatures(features, sample.stats = NA)
}
\arguments{
\item{features}{A dataframe of features}

\item{sample.stats}{If samplle.stats is provided it will normalise features given those sample stats}
}
\value{
A list with two components. Features and sample stats
normalises continous features
}
\description{
Seperate Features into continuous and categorical
}

